requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
asyncio-extras==1.3.2
asyncpraw==7.4.0
asyncprawcore==2.3.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
comtypes==1.1.10
discord-py-slash-command==2.3.2
discord.py==1.7.3
idna==2.10
MouseInfo==0.1.3
multidict==5.1.0
youtube-dl==2021.6.6
Pillow==8.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyttsx3==2.90
PyTweening==1.0.3
requests==2.25.1
simpledemotivators==2.0.1
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
yarl==1.6.3
